chore: functional tests instead of github issues
