--- test/github-issues/11563/entity/TestEntity.ts
@@ -1,10 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column } from "../../../../src/index"
-
-@Entity()
-export class TestEntity {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column({ type: "varchar", length: 100 })
-    name: string
-}
